updates tickets and apps
TICKETS PLATFORM CHANGES

POINT 1: The "REGISTER" was removed from the login, and the switch (see password) was adjusted with the password input so that it is not seen as a separate component.
POINT 2: The "Close ticket" icon was changed.
POINT 3: Push notifications are pending deployment. (Marcos)
POINT 4: The size of the response is different from the size of the text editor, because the text editor component is not formatted. Unlike the response where the size was increased at your request in a previous presentation.
POINT 5: The close ticket button was removed from the gear.
POINT 6: How static?
POINT 7: Pending (Marcos)
POINT 8: The confirmation for sending responses is removed.
POINT 9: The preview of the files is added and they are removed when you click on them